+++ crates/sotf-player/src/controllers/plugin/set.rs
@@ -410,6 +410,8 @@ mod tests {
             max_filters: 10,
             tdf2: false,
             topology: 0.0,
+            auto_gain_enabled: false,
+            oversampling: 1.0,
         };
         let mut changed = false;
 
@@ -437,6 +439,8 @@ mod tests {
             max_filters: 10,
             tdf2: false,
             topology: 0.0,
+            auto_gain_enabled: false,
+            oversampling: 1.0,
         };
         let mut changed = false;
 
@@ -458,6 +462,8 @@ mod tests {
             max_filters: 10,
             tdf2: false,
             topology: 0.0,
+            auto_gain_enabled: false,
+            oversampling: 1.0,
         };
         let mut changed = false;
 
@@ -486,6 +492,8 @@ mod tests {
             max_filters: 10,
             tdf2: false,
             topology: 0.0,
+            auto_gain_enabled: false,
+            oversampling: 1.0,
         };
         let mut changed = false;
 
@@ -506,6 +514,8 @@ mod tests {
             max_filters: 10,
             tdf2: false,
             topology: 0.0,
+            auto_gain_enabled: false,
+            oversampling: 1.0,
         };
         let mut changed = false;
 
@@ -530,6 +540,8 @@ mod tests {
             max_filters: 10,
             tdf2: false,
             topology: 0.0,
+            auto_gain_enabled: false,
+            oversampling: 1.0,
         };
         let mut changed = false;